remove unused classes and rename
